Remove the unsafe_indexing feature
I don't really feel like this feature is needed since performance
already seems fine. I also feel like this should be always on because
its been verified to be correct, or, if we're not sure its correct, it
shouldn't exist. So, since I don't plan on verifying it, I removed it
instead.