Use a table for the generator poly instead of calculating it
The originally library calculated the generator poly at runtime - and so
we kept that until now. For the original library, in GF(2^8), storing
the entire table of all generator polys would take around 32k which is
not insignificant. In GF(2^5), however, it only requires around 512 byts
which is much more manageable.

Storing the polynomials as a constant will also allow us to simplify our
interface by not requiring the construction of an Encoder or Decoder
object.

In order to help keep down binary sizes, we're exposing two interfaces:

* The first is the encode() function which takes the data to encode as
  well as the number of ecc symbols. In then looks up one of the
  pre-computed Encoders and uses that. This method is easy to use - but
  the downside is that it may bring in a bunch of precomputed tables
  that aren't actually needed.

* The second is that const instances of the Encoder struct are provided
  for every choice of ecc symbols - 0 to 30. A client that is worried
  about binary size, may be able to use one of these directly in order
  to reduce the size of the resulting binary by up to about 1,000 bytes
  (in my limited testing). Whether or not this interface has any
  advantage seems to be highly target dependent - but at least for WASM
  it does seem to reduce binary size.