Rework the Error handling interface
Previously we had an EncodeError and a DecodeError enum. Some issues:

* They conflated use-based error where the client code did something it
  shouldn't with actual data errors. This isn't necessarily bad - but,
  I'd prefer to keep these separated since I think the use-based errors
  are hard to handle and make the interface more verbose / harder to
  change.

* Not all error cases where handled - things like incorrect ECC sizes,
  for example.

So, we switch to new error interfaces with two error types: UsageError
and CorrectionError.

A UsageError indicates that an interface was used incorrectly. This
should never occur except in incorrectly written calling code. Printing
out the UsageError will give a description of the mistake.

A significant question is: if invalid input quintets are passed in, is
that a usage error? For something like a Base64 encoding crate - the
answer would almost certainly be no - its an error with the input data.
However, for this crate, I decided to count it as a UsageError. My
reason for doing so is that I can't think of a reasonable way where
client code could end up with an array of quintets values that it wants
to pass to this crate that it hasn't yet validated. This differs from
something like a Base64 crate where it would be more reasonable to pass
in raw, unvalidated external input.

A CorrectionError contains two variants - a UsageError (just like above)
and TooManyErrors. The latter case indicates that a message that was
passed in was invalid and could not be corrected.